requirements.txt
torch==2.1.2
pyyaml==6.0.1
numpy==1.26.4
scipy==1.12.0
tqdm==4.66.1
wandb==0.16.2